fix(gpu-bfs): uint64 emitter counter + fused last-level count for dense queries
Two correctness fixes for large/dense queries on the BFS path, found by
cross-scale regression on Amazon (6v-11v) vs CPU batch_all golden.

P1a — uint64 partial-match counter. The per-layer emitter counter (d_count_
and every expand kernel's out_count) was uint32. A single dense expand layer
can emit more than 2^32 candidates (10v Q_3, 11v Q_1), so atomicAdd wrapped and
the host read a garbage count (10v Q_3 reported 4,294,974,963 ~= 2^32 instead
of 1,315,908,335). Promotes d_count_, the four writer kernels' out_count
params, warp_agg_inc, and all host reset/read paths to 64-bit. cur_count /
total_out / chunk stay 32-bit (bounded by MAX_BUF_MATCHES). Fixes 10v Q_3 and
11v Q_1 to exact CPU values.

P1b — fused expand+count for the last two depths. At depth Q-2 the BFS used to
materialise the depth-(Q-1) layer (the single largest layer, e.g. 277M pms for
9v Q_1) into the ping-pong buffer and then run a separate count kernel over it.
The new bfs_expand_count_kernel extends each depth-(Q-2) candidate in-register
straight to depth Q-1 and counts, so the largest layer is never written to
global memory. This removes the buffer-overflow flush that was silently
dropping matches: 9v Q_1 goes from 1,839,971,202 (non-deterministic, lost ~82)
to the exact, deterministic 1,839,971,284, and 9v Q_3 drops from ~16s to ~1.3s
by avoiding the overflow path. Versioned mode keeps the explicit two-kernel
flow (timestamp logic). Non-versioned only; guarded by Q>=3.

Regression (Amazon full stream, gpu_bfs vs CPU batch_all): 6v/7v/9v/10v
Q_1/Q_3 all match exactly and deterministically.

+++ ParaCOSM/CSM/core/gpu/gpu_bfs_search.h
@@ -101,7 +101,7 @@ private:
     static constexpr size_t MAX_BUF_MATCHES = 400'000'000;
 
     // Counters
-    uint32_t* d_count_ = nullptr;
+    unsigned long long* d_count_ = nullptr;  // logical emitter count (64-bit: dense layers exceed 2^32)
     uint64_t* d_result_ = nullptr;
 
     // Edge input buffers
